update to compressed data

+++ variables/variables_building.py
@@ -68,7 +68,7 @@ def popden(buildings, zones):
 
 @orca.column('buildings', cache=True, cache_scope='iteration')
 def building_sqft(buildings):
-    return buildings.non_residential_sqft + buildings.sqft_per_unit * buildings.residential_units
+    return buildings.non_residential_sqft.astype(int) + buildings.sqft_per_unit * buildings.residential_units
 
 
 @orca.column('buildings', cache=True, cache_scope='iteration')
